moved ssi catalog ingestion base stage to ingest/base as this could be useful for other ingestion stages. also fixed some column name inconsistancies in the balrog stage
