+++ txpipe/ingest/ssi.py
@@ -7,6 +7,7 @@ from ..utils import (
     nanojansky_to_mag_ab,
     nanojansky_err_to_mag_ab,
 )
+from .base import TXIngestCatalogBase
 import numpy as np
 import warnings
 
@@ -16,126 +17,7 @@ column_names = {
     "coord_dec": "dec",
 }
 
-class TXIngestSSI(PipelineStage):
-    """
-    Base-Class for ingesting SSI catalogs
-
-    These can be GCR catalogs, or precursor surveys
-    """
-
-    name = "TXIngestSSI"
-
-    def create_photometry(self, input_name, output_name, column_names, dummy_columns):
-        """
-        Create and save photometry dataset from an input catalog.
-
-        This method sets up the output HDF5 file structure and processes photometry
-        data from the input catalog. It maps the input columns to the desired output
-        column names, creates datasets for photometric magnitudes and their errors,
-        and iterates over chunks of data to save them in the output file.
-
-        Parameters
-        ----------
-        input_name : str
-            name of (FITS) input
-
-        output_name : str
-            name of (HDF5) output
-
-        column_names : dict
-            A dictionary mapping the input column names to the desired output column names.
-            Keys are input column names, and values are the corresponding output names.
-
-        dummy_columns : dict
-            A dictionary of columns to be added to the output with fixed values.
-            Keys are the names of the dummy columns, and values are the constant values
-            to fill those columns.
-
-        """
-        # get some basic onfo about the input file
-        f = self.open_input(input_name)
-        n = f[1].get_nrows()
-        dtypes = f[1].get_rec_dtype()[0]
-        f.close()
-
-        chunk_rows = self.config["chunk_rows"]
-
-        # set up the output file columns 
-        output, g = self.setup_output(output_name, column_names, dtypes, n)
-
-        # iterate over the input file and save to the output columns
-        self.add_columns(g, input_name, column_names, chunk_rows, n)
-
-        # Set up any dummy columns with sentinal values
-        # that were not in the original files
-        for col_name in dummy_columns.keys():
-            g.create_dataset(col_name, data=np.full(n, dummy_columns[col_name]))
-
-        output.close()
-
-    def setup_output(self, output_name, column_names, dtypes, n):
-        """
-        Set up the output HDF5 file structure.
-
-        Parameters
-        ----------
-        output_name : str
-            The name of the output HDF5 file.
-
-        column_names : dict
-            dict of column names to include in the output file.
-
-        dtypes : dict
-            A dictionary mapping column names to their corresponding data types.
-
-        n : int
-            The total number of rows in the dataset.
-
-        Returns
-        -------
-        tuple
-            A tuple containing the open HDF5 output file object and the "photometry" group.
-        """
-        cols = list(column_names.keys())
-        output = self.open_output(output_name)
-        g = output.create_group("photometry")
-        for col in cols:
-            dtype = dtypes[col]
-            g.create_dataset(column_names[col], (n,), dtype=dtype)
-        
-        return output, g
-
-    def add_columns(self, g, input_name, column_names, chunk_rows, n):
-        """
-        Add data to the HDF5 output file in chunks.
-
-        This method reads chunks of data from the input file and writes them
-        to the corresponding datasets in the output file.
-
-        Parameters
-        ----------
-        g : h5py.Group
-            The HDF5 group where data will be written.
-
-        input_name : str
-            The name of the input file (e.g., a FITS file).
-
-        column_names : dict
-            Dict of column names to read from the input file.
-
-        chunk_rows : int
-            Number of rows to process in each chunk.
-
-        n : int
-            Total number of rows in the dataset.
-        """
-        cols = list(column_names.keys())
-        for s, e, data in self.iterate_fits(input_name, 1, cols, chunk_rows):
-            print(s, e, n)
-            for col in cols:
-                g[column_names[col]][s:e] = data[col]
-
-class TXIngestSSIGCR(TXIngestSSI):
+class TXIngestSSIGCR(TXIngestCatalogBase):
     """
     Ingest SSI catalogs using GCR
 
@@ -399,7 +281,7 @@ class TXMatchSSI(PipelineStage):
         outfile.close()
         return
 
-class TXIngestSSIDESBalrog(TXIngestSSI):
+class TXIngestSSIDESBalrog(TXIngestCatalogBase):
     """
     Base-stage for ingesting a DES SSI catalog AKA "Balrog"
     """
@@ -451,7 +333,7 @@ class TXIngestSSIDESBalrog(TXIngestSSI):
                     # "meas_EXTENDED_CLASS_SOF" is (0 or 1) for star, (2 or 3) for galaxy
                     # extendedness is 0 for star, 1 for galaxy
                     extendedness = np.where((data[col] == 2) | (data[col] == 3), 1, 0) 
-                    g[column_names[col]][s:e] = extendedness
+                    g["extendedness"][s:e] = extendedness
 
 
 class TXIngestSSIMatchedDESBalrog(TXIngestSSIDESBalrog):
@@ -491,13 +373,13 @@ class TXIngestSSIMatchedDESBalrog(TXIngestSSIDESBalrog):
         }
         dummy_columns = {
             "redshift_true": 10.0,
-            "meas_err_g":-99.,
-            "meas_err_r":-99.,
-            "meas_err_i":-99.,
-            "meas_err_z":-99.,
+            "mag_err_g":-99.,
+            "mag_err_r":-99.,
+            "mag_err_i":-99.,
+            "mag_err_z":-99.,
         }
 
-        self.create_photometry(
+        self.create_h5_from_fits(
             "balrog_matched_catalog",
             "matched_ssi_photometry_catalog",
             column_names,
@@ -537,7 +419,7 @@ class TXIngestSSIDetectionDESBalrog(TXIngestSSIDESBalrog):
             "meas_EXTENDED_CLASS_SOF": "meas_EXTENDED_CLASS_SOF",  # star galaxy separator
         }
 
-        self.create_photometry(
+        self.create_h5_from_fits(
             "balrog_detection_catalog", "injection_catalog", column_names_inj, {}
         )
 
@@ -555,6 +437,6 @@ class TXIngestSSIDetectionDESBalrog(TXIngestSSIDESBalrog):
             "match_flag_2.0_asec": "match_flag_2.0_asec",
         }
 
-        self.create_photometry(
+        self.create_h5_from_fits(
             "balrog_detection_catalog", "ssi_detection_catalog", column_names_det, {}
         )
